flask/requirements.txt
# requirements.txt

flask
werkzeug
paho-mqtt
python-dotenv
requests
chromadb==0.4.24
langchain-community
langchain-deepseek
numpy<2.0
pypdf
sentence-transformers
ultralytics
torch
pandas
pytz
schedule